# perf: optimizing functions