--- a/index.js
+++ b/index.js
@@ -137,8 +137,8 @@ export function uint8ArrayToBase64(array, { urlSafe = false } = {}) {
 	let base64;
 
 	if (array.length < MAX_BLOCK_SIZE) {
-		// Required as `btoa` and `atob` don't properly support Unicode: https://developer.mozilla.org/en-US/docs/Glossary/Base64#the_unicode_problem
-		base64 = globalThis.btoa(String.fromCodePoint(...array));
+	// Required as `btoa` and `atob` don't properly support Unicode: https://developer.mozilla.org/en-US/docs/Glossary/Base64#the_unicode_problem
+	base64 = globalThis.btoa(String.fromCodePoint.apply(this, array));
 	} else {
 		base64 = '';
 		for (let value of array) {