--- a/package.json
+++ b/package.json
@@ -48,6 +48,7 @@
 	"devDependencies": {
 		"ava": "^5.3.1",
 		"typescript": "^5.2.2",
-		"xo": "^0.56.0"
+		"xo": "^0.56.0",
+		"benchmark": "2.1.4"
 	}
 }