@sindresorhus 2023-11-08
I think you switched the benchmark results, as otherwise, everything got slower.