@sindresorhus 2023-11-08
Thank you :)